Correction bug du sac-a-dos Modifications qui seront validées : 	modifié:         Include/KnapSack.h 	modifié:         Output_ColumnGeneration/GAP-a05100.dat 	modifié:         Run 	modifié:         Src/ColumnGeneration.cpp 	modifié:         Src/KnapSack.cpp 	modifié:         Src/ModelCompact.cpp 	modifié:         Src/ModelMaitre.cpp

+++ Include/KnapSack.h
@@ -22,7 +22,7 @@
  *              (Assume that it is initialize with n elements !)
  * @return: the optimal value of the knapsack
  */
-int knapSack(int W, std::vector<int> wt, std::vector<int> val, int n, IloNumArray oSol);
+int knapSack(double W, std::vector<double> wt, std::vector<double> val, int n, IloNumArray oSol);
 
 #endif
